+++ data-service-base/src/main/java/org/fugerit/java/dsb/DataService.java
@@ -25,5 +25,19 @@ public interface DataService {
 	 * @throws IOException	if any I/O issue arises
 	 */
 	String save( InputStream data ) throws IOException;
+
+	/**
+	 * <p>Save data and assign an ID to it.</p>
+	 *
+	 * <p>Default implementation of this method will invoke <code>this.save( data )</code></p>
+	 *
+	 * @param data				the data to be saved
+	 * @param resourceName		resource name to be saved
+	 * @return					the ID assigned to saved data
+	 * @throws IOException	if any I/O issue arises
+	 */
+	default String save( InputStream data, String resourceName ) throws IOException {
+		return this.save( data );
+	}
 	
 }